Make tests and module conditional on if on campus
Problem with working with MIT ldap server is that in order to not
require authentication, it limits access to the MIT network. Thus it's
hard to develop an application that depends MIT::LDAP::Search as a
network connection is required to connect via VPN.  If not on campus,
the Ldaptic :search and :find methods will return empty stubs of what
they would normally return.