Restructure how IP is checked and how Ldaptic is included
Check actual IP address and not hostname to determine if we're on
campus.  Hostname is not always set by certain networks.

Add MIT::LDAP.connect! that checks if on campus and includes
Ldaptic::Module into singleton class if so.

This provides a cleaner, safer way to load dependencies. Any tools
depending on this Gem should use MIT.on_campus? or check if
MIT::LDAP.connection is nil before making queries.